+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+* **Resolver** Changed the `OrderForm` query to parse the integer prices to float.
+
 ## [2.3.2] - 2018-04-05
 
 ### Fixed